Fixed Room Inconsistency
• Room values were being stored as bytes instead of integer values. Since the value to save was not possible to cast as a byte, the number assigned to objects was incorrect in the JSON files.
• Updated JSON files.

+++ KatAMEntity.cs
@@ -20,7 +20,7 @@ public class Entity {
     public byte Behavior { get; set; }
     public byte Speed { get; set; }
     public byte[] Properties { get; set; } = new byte[18];
-    public byte Room { get; set; }
+    public int Room { get; set; }
 
     public Entity() {}
